+++ z-paging/pages/index/index.vue
@@ -1,6 +1,6 @@
 <template>
 	<view class="content">
-		<z-paging ref="paging" fixed refresherOnly @onRefresh="onRefresh" :refresher-threshold="60">
+		<z-paging ref="paging" refresherOnly @onRefresh="onRefresh" :refresher-threshold="60">
 			<view slot="refresher" style="height: 100%;display: flex;justify-content: center;align-items: center;">
 				<image style="width: 300rpx;height: 60rpx;" src="../../static/logo_loading.gif"></image>
 			</view>