+++ z-paging/pages/sticky-and-scroll-tab-demo/sticky-and-scroll-tab-demo.vue
@@ -1,7 +1,7 @@
 <!-- 滑动切换选项卡+吸顶演示（待完善） -->
 <template>
 	<view class="content">
-		<z-paging ref="paging" @scroll="scroll" :show-console-error="false" :hide-empty-view="true"
+		<z-paging ref="paging" @scroll="scroll" :fixed="false" :show-console-error="false" :hide-empty-view="true"
 			refresher-threshold="160rpx" :refresher-status.sync="refresherStatus" @query="queryList"
 			:paging-content-style="{height:'calc(100% + 260rpx)'}">
 			<!-- 自定义下拉刷新view -->