+++ z-paging/pages/sticky-demo/sticky-demo.vue
@@ -2,7 +2,7 @@
 <template>
 	<view class="content">
 		<!-- 此处为了让reload时不自动滚动到顶部，需要设置auto-clean-list-when-reload和auto-scroll-to-top-when-reload为false，即在reload时关闭自动清空数组和自动滚动到顶部 -->
-		<z-paging ref="paging" v-model="dataList" fixed :auto-clean-list-when-reload="false"
+		<z-paging ref="paging" v-model="dataList" :auto-clean-list-when-reload="false"
 			:auto-scroll-to-top-when-reload="false" refresher-threshold="160rpx"
 			:refresher-status.sync="refresherStatus" @query="queryList">
 			<view class="banner-view" style="height: 250rpx;">