+++ z-paging/uni_modules/z-paging/components/z-paging-empty-view/z-paging-empty-view.vue
@@ -121,7 +121,7 @@
 <style scoped>
 	.zp-container {
 		/* #ifndef APP-NVUE */
-		position: fixed;
+		position: absolute;
 		top: 0;
 		left: 0;
 		width: 100%;