+++ z-paging/uni_modules/z-paging/components/z-paging/z-paging.vue
@@ -4,8 +4,8 @@
   / /_____| |_) | (_| | (_| | | | | | (_| |
  /___|    | .__/ \__,_|\__, |_|_| |_|\__, |
           |_|          |___/         |___/ 
-V1.8.9
-by ZXLee 2021-06-30
+V1.9.0
+by ZXLee 2021-07-04
 -- >
 <!-- API文档地址：http://z-paging.com -->
 <!-- github地址:https://github.com/SmileZXLee/uni-z-paging -->